smooth

+++ Assets/Scripts/GPSLoader.cs
@@ -23,8 +23,8 @@ public class GPSLoader : MonoBehaviour {
 		} else {
 			while(true){
 				yield return new WaitForSeconds(0.5f);
-				Utils.lat =  (double)Mathf.Lerp((float)Utils.lat,Input.location.lastData.latitude,0.9f);
-				Utils.lang = (double)Mathf.Lerp((float)Utils.lang,Input.location.lastData.longitude,0.9f);
+				Utils.lat =  Mathf.Lerp((float)Utils.lat,Input.location.lastData.latitude,0.9f);
+				Utils.lang = Mathf.Lerp((float)Utils.lang,Input.location.lastData.longitude,0.9f);
 				//Utils.alti = Input.location.lastData.altitude;
 
 				 Debug.Log("Location: " +
